+++ src/main/java/com/wholeseeds/mindle/domain/complaint/Service/ComplaintService.java
@@ -4,7 +4,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.wholeseeds.mindle.common.service.NcpObjectStorageService;
 import com.wholeseeds.mindle.domain.complaint.dto.SaveComplaintRequestDto;
 import com.wholeseeds.mindle.domain.complaint.entity.Category;
 import com.wholeseeds.mindle.domain.complaint.entity.Complaint;
@@ -27,6 +26,7 @@ import com.wholeseeds.mindle.domain.member.exception.MemberNotFoundException;
 import com.wholeseeds.mindle.domain.member.repository.MemberRepository;
 import com.wholeseeds.mindle.domain.place.entity.Place;
 import com.wholeseeds.mindle.domain.place.repository.PlaceRepository;
+import com.wholeseeds.mindle.infra.Service.NcpObjectStorageService;
 
 import jakarta.annotation.Nullable;
 import lombok.RequiredArgsConstructor;
